[racingcar]refactor: set constant
