+++ src/main/java/racingcar/car/Cars.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.car;
 
 import java.util.Arrays;
