+++ src/test/java/racingcar/RacingCarTest.java
@@ -4,6 +4,8 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
+import racingcar.car.Car;
+import racingcar.car.Cars;
 
 import static org.assertj.core.api.Assertions.*;
